requirements-dev.txt
# Dependências de desenvolvimento

# Dependências principais
-r requirements.txt

# Debug e desenvolvimento
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testes
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
flake8==6.1.0
black==23.12.0
isort==5.13.2

# Documentação
sphinx==7.2.6


